Fix BoolToObjectConverter.ConvertBack
`targetType` in a `ConvertBack` operation of this class will always be
something like `bool`, so `TrueValue` cannot be converted to that, it
should be converted to the same type as the `value` parameter, since
they should match.

A check is added using `targetType` to verify that the resulting `bool`
can be assigned to the target.